Models/CoMet-SN/requirements.txt
#click==8.3.0
click
matplotlib==3.7.5
numpy==1.22.4
opencv_python==4.11.0.86
pandas==1.5.2
#Pillow==12.0.0
Pillow
psutil==5.9.8
scikit_learn==1.3.2
#scipy==1.16.2
scipy
#skimage==0.0
scikit-image
timm==1.0.21
#torch==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.67.1
